# app/Http/Controllers/UserController.php
<?php

namespace App\Http\Controllers;

use App\Models\User;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Hash;
use App\Http\Requests\UserRequest;

class UserController extends Controller
{
    /**
     * @OA\Get(
     *     path="/api/users",
     *     tags={"Users"},
     *     summary="Get all users",
     *     operationId="getUsers",
     *     @OA\Response(
     *         response=200,
     *         description="successful operation",
     *     ),
     *     @OA\Response(
     *         response=400,
     *         description="Invalid status value"
     *     ),
     *     security={{"bearerToken":{}}}
     * )
     */
    public function getUsers()
    {
        return User::all();
    }

    /**
     * Register a user.
     *
     * @OA\Post(
     *     path="/api/register",
     *     tags={"Users"},
     *     operationId="register",
     *     @OA\Response(
     *         response=405,
     *         description="Validation exception"
     *     ),
     *     @OA\RequestBody(
     *         description="Input data format",
     *         @OA\MediaType(
     *             mediaType="application/x-www-form-urlencoded",
     *             @OA\Schema(
     *                 type="object",
     *                 required={
     *                  "first_name",
     *                  "last_name",
     *                  "email",
     *                  "password"
     *                  },
     *                 @OA\Property(
     *                     property="first_name",
     *                     description="Insert the username",
     *                     type="string",
     *                 ),
     *                 @OA\Property(
     *                     property="last_name",
     *                     description="Insert the last name",
     *                     type="string"
     *                 ),
     *                 @OA\Property(
     *                     property="email",
     *                     description="Insert the email",
     *                     type="string"
     *                 ),
     *                 @OA\Property(
     *                     property="password",
     *                     description="Insert the password",
     *                     type="string"
     *                 )
     *             )
     *         )
     *     )
     * )
     */
    public function register(UserRequest $request)
    {
        try{

            $user = User::create([
                "first_name" => $request->input("first_name"),
                "last_name" => $request->input("last_name"),
                "email" => $request->input("email"),
                "password" => Hash::make(trim($request->input("password")))
            ]);

            $token = $user->createToken('user_token')->plainTextToken;

            return response()->json([
                "status" => "Success",
                "data" => $user,
                "token" => $token,
                "mensaje" => "Success operation"
            ], 200);

        }catch(\Exception $e){
            return response()->json([
                "error" => $e->getMessage(),
                "message" => "Something went wrong in register method"
            ]);
        }

    }

    /**
     * Display the specified resource.
     */
    public function show(string $id)
    {
        $data = User::find($id);

        if (isset($data)) {
            return response()->json([
                "data" => $data,
                "status" => "Success",
                "mensaje" => "Success operation"
            ]);
        } else {
            return response()->json([
                "data" => null,
                "status" => "error",
                "mensaje" => "The element no exist"
            ]);
        }
    }

    /**
     * Update the specified resource in storage.
     */
    public function update(Request $request, string $id)
    {
        $data = User::find($id);

        if (isset($data)) {
            $data->first_name = $request->first_name;
            $data->last_name = $request->last_name;
            $data->email = $request->email;
            $data->password = Hash::make(trim($request->password));

            if ($data->save()) {
                return response()->json([
                    "data" => $data,
                    "status" => "Success",
                    "mensaje" => "Success operation"
                ]);
            } else {
                return response()->json([
                    "data" => null,
                    "status" => "error",
                    "mensaje" => "Something went wrong in update method"
                ]);
            }
        } else {
            return response()->json([
                "data" => null,
                "status" => "error",
                "mensaje" => "The element no exist"
            ]);
        }
    }

    /**
     * Remove the specified resource from storage.
     */
    public function destroy(string $id)
    {
        $data = User::find($id);

        if (isset($data)) {
            if (User::destroy($id)) {
                return response()->json([
                    "data" => $data,
                    "status" => "Success",
                    "mensaje" => "Success operation"
                ]);
            } else {
                return response()->json([
                    "data" => null,
                    "status" => "error",
                    "mensaje" => "Something went wrong in delete method"
                ]);
            }
        } else {
            return response()->json([
                "data" => null,
                "status" => "error",
                "mensaje" => "The element no exits"
            ]);
        }
    }
}